+++ data/src/main/java/com/tht/tht/data/di/UseCaseModule.kt
@@ -8,7 +8,7 @@ import com.tht.tht.domain.signup.repository.LocationRepository
 import com.tht.tht.domain.signup.repository.SignupRepository
 import com.tht.tht.domain.signup.usecase.*
 import com.tht.tht.domain.token.repository.TokenRepository
-import com.tht.tht.domain.token.usecase.RequestFcmTokenLoginUseCase
+import com.tht.tht.domain.login.usecase.RequestFcmTokenLoginUseCase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
